+++ core/src/main.rs
@@ -75,24 +75,24 @@ fn main() {
     }
 
     // Command Line Interface for looking up Compounds
-    loop {
-        println!("Enter your input, or type 'quit' to exit");
-        let mut user_input = String::new();
-        match io::stdin().read_line(&mut user_input) {
-            Ok(_) =>  {
-                if user_input.trim().to_lowercase() == "quit" || user_input.trim().to_lowercase() == "'quit'" {
-                    break
-                }
-            },
-            Err(_) => println!("Error"),
-        }
-        let response = compound_trees.get(&user_input.trim().to_lowercase());
+    // loop {
+    //     println!("Enter your input, or type 'quit' to exit");
+    //     let mut user_input = String::new();
+    //     match io::stdin().read_line(&mut user_input) {
+    //         Ok(_) =>  {
+    //             if user_input.trim().to_lowercase() == "quit" || user_input.trim().to_lowercase() == "'quit'" {
+    //                 break
+    //             }
+    //         },
+    //         Err(_) => println!("Error"),
+    //     }
+    //     let response = compound_trees.get(&user_input.trim().to_lowercase());
 
-        match response {
-            Some(x) =>  { x.print_dispenser_format() },
-            None => { 
-                println!("{} is not a valid Compound!", user_input.trim());
-            }
-        }
-    }
+    //     match response {
+    //         Some(x) =>  { x.print_dispenser_format() },
+    //         None => { 
+    //             println!("{} is not a valid Compound!", user_input.trim());
+    //         }
+    //     }
+    // }
 }
